名前空間のネストが深いので浅くするのとクラス名を変更
