+++ app/controllers/articles_controller.rb
@@ -32,7 +32,7 @@ class ArticlesController < ApplicationController
     @article.user = current_user if @article.user.nil?
     set_wip
     if @article.save
-      Ogp::Image::AttachmentProcessor.fit_to_size(@article.thumbnail) if @article.thumbnail.attached?
+      Ogp::ImageProcessor.fit_to_size(@article.thumbnail) if @article.thumbnail.attached?
 
       redirect_to redirect_url(@article), notice: notice_message(@article)
     else
@@ -43,7 +43,7 @@ class ArticlesController < ApplicationController
   def update
     set_wip
     if @article.update(article_params)
-      Ogp::Image::AttachmentProcessor.fit_to_size(@article.thumbnail) if @article.thumbnail.attached?
+      Ogp::ImageProcessor.fit_to_size(@article.thumbnail) if @article.thumbnail.attached?
 
       redirect_to redirect_url(@article), notice: notice_message(@article)
     else
